Update the way we collect perf metric for ACTN (function call contrl path).
Also, change perf metric data type from "int" to "long".

+++ ChpProcGenerator.h
@@ -1,7 +1,7 @@
 #ifndef DFLOWMAP_CHPPROCGENERATOR_H
 #define DFLOWMAP_CHPPROCGENERATOR_H
 
-#include <string.h>
+#include <cstring>
 #include <algorithm>
 #include <fstream>
 #include "common.h"
@@ -15,26 +15,26 @@ public:
   void
   createFULib(FILE *libFp, FILE *confFp, const char *procName, const char *calc,
               const char *def, const char *outSend, int numArgs,
-              int numOuts, int numRes, const char *instance, int *metric,
+              int numOuts, int numRes, const char *instance, long *metric,
               IntVec &boolRes);
 
   void
   createMerge(FILE *libFp, FILE *confFp, const char *procName, const char *instance,
-              int *metric, int numInputs);
+              long *metric, int numInputs);
 
   void
   createSplit(FILE *libFp, FILE *confFp, const char *procName, const char *instance,
-              int *metric, int numOutputs);
+              long *metric, int numOutputs);
 
-  void createSource(FILE *libFp, FILE *confFp, const char *instance, int *metric);
+  void createSource(FILE *libFp, FILE *confFp, const char *instance, long *metric);
 
-  void createInit(FILE *libFp, FILE *confFp, const char *instance, int *metric);
+  void createInit(FILE *libFp, FILE *confFp, const char *instance, long *metric);
 
-  void createBuff(FILE *libFp, FILE *confFp, const char *instance, int *metric);
+  void createBuff(FILE *libFp, FILE *confFp, const char *instance, long *metric);
 
-  void createSink(FILE *libFp, FILE *confFp, const char *instance, int *metric);
+  void createSink(FILE *libFp, FILE *confFp, const char *instance, long *metric);
 
-  void createCopy(FILE *libFp, FILE *confFp, const char *instance, int *metric);
+  void createCopy(FILE *libFp, FILE *confFp, const char *instance, long *metric);
 
 private:
   const char *processes[MAX_PROCESSES];
